requirements.txt
aiogram==3.*
aiosmtplib>=2.0
python-dotenv>=1.0
aiohttp>=3.9
